Move the test code and resources to standard Gradle locations.
Also remove JKidlTest, contains no tests
