+++ src/test/resources/us/kbase/test/kidl/spec.1.properties
