+++ src/test/resources/us/kbase/test/kidl/spec.10.properties
