+++ src/test/resources/us/kbase/test/kidl/spec.11.properties
