+++ src/test/resources/us/kbase/test/kidl/spec.12.properties
