+++ src/test/resources/us/kbase/test/kidl/spec.13.properties
