+++ src/test/resources/us/kbase/test/kidl/spec.14.properties
