+++ src/test/resources/us/kbase/test/kidl/spec.15.properties
