+++ src/test/resources/us/kbase/test/kidl/spec.16.properties
