+++ src/test/resources/us/kbase/test/kidl/spec.17.properties
