+++ src/test/resources/us/kbase/test/kidl/spec.18.properties
