+++ src/test/resources/us/kbase/test/kidl/spec.19.properties
