+++ src/test/resources/us/kbase/test/kidl/spec.2.properties
