+++ src/test/resources/us/kbase/test/kidl/spec.20.properties
