+++ src/test/resources/us/kbase/test/kidl/spec.21.properties
