+++ src/test/resources/us/kbase/test/kidl/spec.22.properties
