+++ src/test/resources/us/kbase/test/kidl/spec.3.properties
