+++ src/test/resources/us/kbase/test/kidl/spec.4.properties
