+++ src/test/resources/us/kbase/test/kidl/spec.5.properties
