+++ src/test/resources/us/kbase/test/kidl/spec.6.properties
