+++ src/test/resources/us/kbase/test/kidl/spec.7.properties
