+++ src/test/resources/us/kbase/test/kidl/spec.8.properties
