+++ src/test/resources/us/kbase/test/kidl/spec.9.properties
